Some rewrites to the rendering engine: -Store paint object for lines. -Use built-in transformations.
git-svn-id: http://dungeon-sketch.googlecode.com/svn/trunk/dungeon-sketch@152 68e44389-5778-9b72-1722-c5411804bbf0

+++ src/com/tbocek/android/combatmap/CombatMap.java
@@ -143,7 +143,7 @@ public final class CombatMap extends Activity {
         }
 
         @Override
-        public void onChooseStrokeWidth(final int width) {
+        public void onChooseStrokeWidth(final float width) {
             mCombatView.setDrawMode();
             mCombatView.setNewLineStrokeWidth(width);
